+++ packages/inputs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @under-control/inputs
 
+## 1.7.17
+
+### Patch Changes
+
+- Fix broken nested validation extract for array
+- Updated dependencies
+  - @under-control/core@1.7.17
+
 ## 1.7.16
 
 ### Patch Changes